n should be an int

+++ wikipedia/track.py
@@ -244,7 +244,7 @@ async def create_users_and_roles(es, params):
 #        except:
 #            pass
 
-    for n, roles in batched(ROLE_IDS, len(ROLE_IDS)/10):
+    for n, roles in batched(ROLE_IDS, int(len(ROLE_IDS)/10)):
         await es.security.put_user(
             username=USERS[n]["username"], params={"roles": roles}
         )
